+++ typescript/sdk/src/consts/multisigIsm.ts
@@ -47,12 +47,13 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
   },
 
   adichain: {
-    threshold: 1,
+    threshold: 2,
     validators: [
       {
         address: '0x4b11a6310bc06300b529b0397683ca3376407eca',
         alias: AW_VALIDATOR_ALIAS,
       },
+      DEFAULT_MITOSIS_VALIDATOR,
     ],
   },
 
@@ -1139,12 +1140,13 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
   },
 
   lazai: {
-    threshold: 1,
+    threshold: 2,
     validators: [
       {
         address: '0x3b00fe3518e739bb978b04d28e1492d8d865d96e',
         alias: AW_VALIDATOR_ALIAS,
       },
+      DEFAULT_MITOSIS_VALIDATOR,
     ],
   },
 
@@ -1304,12 +1306,13 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
   },
 
   megaeth: {
-    threshold: 1,
+    threshold: 2,
     validators: [
       {
         address: '0x051ddac8ecf4bae2532b8b7caa626b5567dab528',
         alias: AW_VALIDATOR_ALIAS,
       },
+      DEFAULT_MITOSIS_VALIDATOR,
     ],
   },
 
@@ -2069,12 +2072,13 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
   },
 
   somnia: {
-    threshold: 1,
+    threshold: 2,
     validators: [
       {
         address: '0xf484907083d32fdc0848bfb998dfdde835e6f9cb',
         alias: AW_VALIDATOR_ALIAS,
       },
+      DEFAULT_MITOSIS_VALIDATOR,
     ],
   },
 
@@ -2184,12 +2188,13 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
   },
 
   stable: {
-    threshold: 1,
+    threshold: 2,
     validators: [
       {
         address: '0x21820baebcd972c769e490415cfee43a894f3c18',
         alias: AW_VALIDATOR_ALIAS,
       },
+      DEFAULT_MITOSIS_VALIDATOR,
     ],
   },