+++ typescript/cosmos-types/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @hyperlane-xyz/cosmos-types
 
+## 19.13.0
+
 ## 19.12.0
 
 ## 19.11.0